Anime & Character Endpoint Completed
