+++ src/common/structs/anime.rs
@@ -1,7 +1,12 @@
 use serde::{Deserialize, Serialize};
 use crate::{
-    utils::{Images, DateRange, Title},
+    utils::{Images, DateRange, Title, Score},
     response::MalCommonResponse,
+    structs::{
+        people::Person,
+        character::Character,
+        forum::ForumTopic,
+    }
 };
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -73,4 +78,71 @@ pub struct AnimeExtended {
     pub explicit_genres: Option<Vec<MalCommonResponse>>,
     pub themes: Option<Vec<MalCommonResponse>>,
     pub demographics: Option<Vec<MalCommonResponse>>,
+}
+
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct PromoVideo {
+    pub title: String,
+    pub trailer: Trailer,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Episode {
+    pub mal_id: i32,
+    pub url: Option<String>,
+    pub title: String,
+    pub episode: Option<String>,
+    pub aired: Option<String>,
+    pub score: Option<f32>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Videos {
+    pub promo: Vec<PromoVideo>,
+    pub episodes: Vec<Episode>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct AnimeCharacter {
+    pub character: Character,
+    pub role: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct StaffMember {
+    pub person: Person,
+    pub positions: Vec<String>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct AnimeThemes {
+    pub openings: Vec<String>,
+    pub endings: Vec<String>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct AnimeForum {
+    pub data: Vec<ForumTopic>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Picture {
+    pub images: Images,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct MoreInfo {
+    pub moreinfo: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct AnimeStatistics {
+    pub watching: i32,
+    pub completed: i32,
+    pub on_hold: i32,
+    pub dropped: i32,
+    pub plan_to_watch: i32,
+    pub total: i32,
+    pub scores: Vec<Score>,
 }
